requirements.txt
absl-py==0.14.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
filelock==3.3.0
flake8==4.0.1
future==0.18.2
gdown==4.0.2
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.4.0
idna==3.2
imageio==2.9.0
isort==4.3.21
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mccabe==0.6.1
numpy==1.20.0
oauthlib==3.1.1
opencv-python==4.5.3.56
Pillow==8.3.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
six==1.16.0
tb-nightly==2.7.0a20211011
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yacs==0.1.8
yapf==0.31.0
pandas==1.3.3
scikit-learn==1.0.1
umap-learn==0.5.2